create as shared component the token component
