+++ apps/web/src/app/shared/components/api-token-generator/index.ts
@@ -1,0 +1,1 @@
+export * from './api-token-generator.component';
