+++ reports/security/consolidated-audit.json
@@ -1,5 +1,5 @@
 {
-  "timestamp": "2025-09-25T20:53:02.133Z",
+  "timestamp": "2025-09-25T23:11:02.590Z",
   "projects": {
     "api": {
       "vulnerabilities": {
